1651 reformat code
