+++ monai/networks/nets/__init__.py
@@ -17,11 +17,10 @@ from .densenet import DenseNet, densenet121, densenet169, densenet201, densenet2
 from .dynunet import DynUNet, DynUnet, Dynunet
 from .fullyconnectednet import FullyConnectedNet, VarFullyConnectedNet
 from .generator import Generator
-from .globalnet import GlobalNet
 from .highresnet import HighResBlock, HighResNet
 from .localnet import LocalNet
 from .regressor import Regressor
-from .regunet import RegUNet
+from .regunet import GlobalNet, RegUNet
 from .segresnet import SegResNet, SegResNetVAE
 from .senet import SENet, se_resnet50, se_resnet101, se_resnet152, se_resnext50_32x4d, se_resnext101_32x4d, senet154
 from .unet import UNet, Unet, unet
